+++ deflect/Server.cpp
@@ -82,7 +82,7 @@ Server::Server( const int port )
         throw std::runtime_error( err.toStdString( ));
     }
 #ifdef DEFLECT_USE_SERVUS
-    _impl->servus.announce( port, boost::lexical_cast< std::string >( port ));
+    _impl->servus.announce( port, "" );
 #endif
 }